+++ docs/js/searchIndex.js
@@ -51,6 +51,11 @@ Search.appendIndex(
             "summary": "Exception\u0020thrown\u0020when\u0020a\u0020class\u0020or\u0020enum\u0020constant\u0020fails\u0020validation.",
             "url": "classes/oihana-reflect-exceptions-ConstantException.html"
         },                {
+            "fqsen": "\\oihana\\reflect\\helpers\\getFunctionInfo\u0028\u0029",
+            "name": "getFunctionInfo",
+            "summary": "Returns\u0020detailed\u0020reflection\u0020information\u0020about\u0020a\u0020given\u0020function\u0020or\u0020method.",
+            "url": "namespaces/oihana-reflect-helpers.html#function_getFunctionInfo"
+        },                {
             "fqsen": "\\oihana\\reflect\\Reflection",
             "name": "Reflection",
             "summary": "High\u002Dlevel\u0020helper\u0020over\u0020PHP\u0027s\u0020Reflection\u0020API\u0020and\u0020a\u0020robust\u0020array\u002Dto\u002Dobject\u0020hydrator.",
@@ -355,6 +360,11 @@ Search.appendIndex(
             "name": "exceptions",
             "summary": "",
             "url": "namespaces/oihana-reflect-exceptions.html"
+        },                {
+            "fqsen": "\\oihana\\reflect\\helpers",
+            "name": "helpers",
+            "summary": "",
+            "url": "namespaces/oihana-reflect-helpers.html"
         },                {
             "fqsen": "\\oihana\\reflect",
             "name": "reflect",