+++ tests/oihana/reflect/helpers/samples/SampleClass.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace tests\oihana\reflect\helpers\samples;
+
+class SampleClass
+{
+    /**
+     * A static test method.
+     */
+    public static function staticMethod(): void {}
+
+    /**
+     * A non-static test method.
+     */
+    public function instanceMethod(): void {}
+}